requirements.txt
Adafruit-Blinka==8.55.0
adafruit-circuitpython-ads1x15==2.4.2
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.8.0
binho-host-adapter==0.1.6
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.1.0
google-auth==2.38.0
google-genai==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
opencv-python==4.11.0.86
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0b1
pydantic_core==2.31.1
pyftdi==0.56.0
pyserial==3.5
PySimpleGUI==5.0.9
python-dotenv==1.0.1
pyusb==1.3.1
requests==2.32.3
RPi.GPIO==0.7.1
rsa==4.9
sniffio==1.3.1
sysv_ipc==1.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
